major rewrite, tool is now pretty much working
extracted files are saved in a .snu file that can be converted to a playable wav file by ToWav.
fixed the way the snr file is parsed to be more correct
included is a renaming tool that extracts the names from the audio bank xml file. this may only be useful for NBA Jam, but who knows?
thanks to Apollo for the help on the .snu file format